+++ rest-api-spec/src/main/resources/rest-api-spec/test/search/20_flattened_stats.yml
