+++ rest-api-spec/src/main/resources/rest-api-spec/test/search/340_flattened.yml
@@ -1,9 +1,10 @@
+setup:
+  - skip:
+      version: " - 7.99.99"
+      reason: "Flattened fields were moved from xpack to core in 7.12."
+
 ---
 "Test exists query on flattened object field":
-  - skip:
-      version: " - 7.2.99"
-      reason: "Flattened fields were implemented in 7.3."
-
   - do:
       indices.create:
         index:  flattened_test
@@ -53,10 +54,6 @@
 
 ---
 "Test query string query on flattened object field":
-  - skip:
-      version: " - 7.2.99"
-      reason: "Flattened fields were implemented in 7.3."
-
   - do:
       indices.create:
         index:  test
@@ -159,9 +156,6 @@
 
 ---
 "Test fields option on flattened object field with include_unmapped":
-  -  skip:
-        version: ' - 7.10.99'
-        reason: support was introduced in 7.11
   - do:
       indices.create:
         index:  test
