--- x-pack/plugin/mapper-flattened/build.gradle
@@ -1,16 +1,0 @@
-apply plugin: 'elasticsearch.esplugin'
-apply plugin: 'elasticsearch.internal-cluster-test'
-
-esplugin {
-  name 'flattened'
-  description 'Module for the flattened field type, which allows JSON objects to be flattened into a single field.'
-  classname 'org.elasticsearch.xpack.flattened.FlattenedMapperPlugin'
-  extendedPlugins = ['x-pack-core']
-}
-archivesBaseName = 'x-pack-flattened'
-
-dependencies {
-  compileOnly project(path: xpackModule('core'))
-  testImplementation(testArtifact(project(xpackModule('core'))))
-}
-
